fix: move scanapi/__main__.py to scanapi/cli.py to debug with pdb works
